pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm[simple]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "Game_Scheduler"
license = "MIT"
dynamic = ["version"]
authors = [
  {name = "Bret McKee", email = "bret.mckee@gmail.com"},
]
description = "Discord game scheduling system with microservices architecture"
requires-python = ">=3.13"
dependencies = [
    # Core Framework
    "discord.py~=2.6.0",
    "audioop-lts~=0.2.0",  # Python 3.13+ compatibility for discord.py voice support
    "fastapi~=0.115.0",
    "slowapi~=0.1.9",
    "uvicorn[standard]~=0.34.0",
    # Database
    "sqlalchemy[asyncio]~=2.0.36",
    "asyncpg~=0.30.0",
    "psycopg2-binary~=2.9.10",
    "alembic~=1.14.0",
    "alembic-utils~=0.8.0",
    # Data Validation
    "pydantic~=2.10.0",
    "pydantic-settings~=2.7.0",
    # Storage & Messaging
    "redis~=5.2.0",
    "aio-pika~=9.5.0",
    "pika~=1.3.0",
    # HTTP Clients
    "httpx~=0.28.0",
    "aiohttp~=3.11.0",
    "python-multipart~=0.0.20",
    "chardet>=3.0.2,<6.0.0",  # requests requires chardet < 6.0.0
    # Security
    "cryptography~=44.0.0",
    "python-jose[cryptography]~=3.3.0",
    # Utilities
    "icalendar~=6.0.0",
    # Observability
    "opentelemetry-api~=1.29.0",
    "opentelemetry-sdk~=1.29.0",
    "opentelemetry-instrumentation-fastapi~=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy~=0.50b0",
    "opentelemetry-instrumentation-asyncpg~=0.50b0",
    "opentelemetry-instrumentation-redis~=0.50b0",
    "opentelemetry-instrumentation-aio-pika~=0.50b0",
    "opentelemetry-exporter-otlp~=1.29.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-timeout>=2.3.0",
    "coverage[toml]>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.15.4",
    "mypy>=1.10.0",
    "diff-cover>=9.2.0",
    "complexipy>=4.2.0",
    "lizard>=1.20.0",
    "pre-commit>=4.0.0",
    "pytest-randomly>=3.16.0",
    "pytest-order>=1.2.0",
]

[project.urls]
Repository = "https://github.com/game-scheduler"
Issues = "https://github.com/game-scheduler/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*", "services*"]
exclude = ["tests*", "frontend*", "docker*", "alembic*", "rabbitmq*", "templates*", "history*", "grafana-alloy*", "env*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py313"
preview = true

[tool.ruff.lint]
select = [
    "E", "F", "I", "N", "W", "B", "C4", "UP",
    "PLR2004", "PLC0415", "C901", "PLR0915",
    "S",      # flake8-bandit security checks
    "ASYNC",  # flake8-async
    "FAST",   # FastAPI-specific
    "T20",    # flake8-print (no print statements in production code)
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TC",     # flake8-type-checking
    "PLE",    # Pylint errors
    "PLW",    # Pylint warnings
    "PLC",    # Pylint conventions
    "ERA",    # eradicate (commented-out code)
    "A",      # flake8-builtins
    "DTZ",    # flake8-datetimez
    "ICN",    # flake8-import-conventions
    "PT",     # flake8-pytest-style
    "G004",   # logging-f-string (enforce lazy logging)
    "EM",     # flake8-errmsg (exception messages)
    "PERF",   # perflint (performance)
    "G",      # flake8-logging-format
    "LOG",    # flake8-logging
    "RUF",    # Ruff-specific rules
    "ARG",    # flake8-unused-arguments
    "ANN",    # flake8-annotations
]
ignore = [
    "S101",    # Allow assert (tests handle separately)
    "RUF029",  # Unnecessary async (false positives in FastAPI/discord.py)
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-statements = 50

[tool.complexipy]
max-complexity-allowed = 15
paths = ["services", "shared", "scripts"]
exclude = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004",  # Magic values (existing)
    "PLR0915",  # Too many statements in test functions
    "S101",     # Assert usage
    "S104",     # Binding to all interfaces in test configs
    "S106",     # Hardcoded passwords in fixtures
    "S105",     # Hardcoded password strings
    "S108",     # Temp file/directory paths in test mocks
    "S110",     # try-except-pass in test fixtures
    "T201",     # Print statements allowed in tests
    "PLC2701",  # Private imports allowed for testing internal functions
    "PLC1901",  # String comparison to empty string in assertions
    "PLC0206",  # Dictionary iteration without .items()
    "SIM102",   # Nested if statements in test helpers
    "SIM105",   # try-except-pass patterns
    "SIM108",   # Ternary operator simplification
    "SIM117",   # Nested with statements in test mocks
    "PT011",    # pytest.raises without match parameter
    "PT012",    # pytest.raises with multiple statements
    "PT018",    # Compound assertions in tests
    "DTZ001",   # datetime without tzinfo in test data
    "ASYNC109", # Async function with timeout parameter
    "ARG001",   # Unused function arguments in fixtures and test helpers
    "ARG002",   # Unused method arguments in test classes
    "ARG004",   # Unused static method arguments in test helpers
    "ARG005",   # Unused lambda arguments in test mocks
    "ANN",      # Type annotations not required in tests
    "C901",     # Complex test functions allowed
    "RUF043",   # Regex metacharacters in pytest.raises match patterns
    "RUF059",   # Unpacked variables not used (common in test assertions)
]
"scripts/**/*.py" = [
    "T201",     # Print statements allowed in CLI scripts
]
[tool.ruff.lint.isort]
known-first-party = ["shared", "services"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
timeout = 30
markers = [
    "integration: Integration tests requiring RabbitMQ, Postgres, Redis",
    "e2e: End-to-end tests requiring Discord bot and full stack",
    "order: Test execution order (used with pytest-order plugin)",
]
addopts = "-m 'not e2e and not integration' --strict-markers --import-mode=importlib"
filterwarnings = [
    "error",
    'ignore::DeprecationWarning:pika.data',
    'ignore::ResourceWarning',
]

[tool.mypy]
python_version = "3.13"
# Enable helpful checks without being overly strict
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
explicit_package_bases = true
namespace_packages = true

# Disable strict checks that cause issues with external libraries
# and legacy code
no_implicit_optional = false
strict_optional = false

# Ignore missing imports from external libraries
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest~=8.3.0",
    "pytest-asyncio~=0.24.0",
    "pytest-cov~=6.0.0",
    "diff-cover~=9.2.0",
    "ruff==0.15.4",
    "mypy~=1.15.0",
    "autocopyright~=1.1.0",
    "pre-commit~=4.0.0",
    "pytest-randomly>=4.0.1",
    "pytest-order>=1.3.0",
]

[tool.coverage.run]
source = ["shared", "services"]
relative_files = true
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "services/init/*",
]

[tool.coverage.report]
precision = 2
skip_empty = true

[tool.duplicate-check]
min-lines = 0

[tool.duplicate-check.overrides]
"tests/**" = 10

[tool.setuptools_scm]
fallback_version = "0.0.0+unknown"
